[utils.ocm] refactor and move models into base file to avoid circular imports
