+++ reconcile/utils/ocm/service_log.py
@@ -1,71 +1,16 @@
+from collections.abc import Generator
 from datetime import (
     datetime,
     timedelta,
 )
-from enum import Enum
-from typing import (
-    Generator,
-    Optional,
+from typing import Optional
+
+from reconcile.utils.ocm.base import (
+    OCMClusterServiceLog,
+    OCMClusterServiceLogCreateModel,
 )
-
-from pydantic import BaseModel
-
 from reconcile.utils.ocm.search_filters import Filter
 from reconcile.utils.ocm_base_client import OCMBaseClient
-
-
-class OCMServiceLogSeverity(str, Enum):
-    """
-    Represents the severity of a service log.
-    """
-
-    Debug = "Debug"
-    Info = "Info"
-    Warning = "Warning"
-    Error = "Error"
-    Fatal = "Fatal"
-
-
-class OCMClusterServiceLogCreateModel(BaseModel):
-    cluster_uuid: str
-    """
-    The cluster UUID is the same as external ID on the OCM cluster_mgmt API
-    """
-
-    service_name: str
-    """
-    The name of the service a service log entry belongs to
-    """
-
-    summary: str
-    """
-    Short summary of the log entry.
-    """
-
-    description: str
-    """
-    Detailed description of the log entry.
-    """
-
-    severity: OCMServiceLogSeverity
-
-
-class OCMClusterServiceLog(OCMClusterServiceLogCreateModel):
-    """
-    Represents a service log entry for a cluster.
-    """
-
-    id: str
-    href: str
-    event_stream_id: str
-    username: str
-
-    cluster_id: str
-
-    timestamp: datetime
-    """
-    The time at which the log entry was created.
-    """
 
 
 def get_service_logs(
